pyproject.toml
[tool.poetry]
name = "bus_tickets_service"
version = "0.1.0"
description = "A comprehensive web application for a company that provides bus transportation services."
authors = ["Olha Kashyrina"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
pydantic-settings = "^2.10.1"
asyncpg = "^0.30.0"
boto3 = "^1.40.26"
psycopg2 = "^2.9.10"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
sqlalchemy-utils = "^0.42.0"
testcontainers = "^4.12.0"
